Add default dictionary and word list files
A default dictionaryFile.txt and wordListFile.txt have been included in
the root directory of the project. Allows the project to be functional
without the need for the user to provide their own dictionary and word
list files.

README.md has been updated to reflect this change.

Fixed typographical errors in program output and documentation.